Code for Journal and AddEntry
I added a addEntry file where I will be doing the insertion to the database table. I also added a touchable opacity button to the journal.js page, where the goal is to navigate to addEntry.js screen. however, button is not function as have not implemented the routing yet.