added a mood tracker
I decided to replace the bookmark entry in our journalEntry table with a mood entry. This is so that users can now track their basic moods along with their entries. I updated the code to also track mood selected. In the UI, when adding the mood tracker, my items were out of alignment, AI helped me fix this so now the emoji appears next to the journal entries.